Add zenml.show()
